Rip out all the old memory limit support (apart from Docker) because this was never correct. The problem is that several of these tools fork multiple processes and so rlimit() and the ulimit hack were enforcing that each process didn't go over the memory limit instead of the desired enforcement that the sum of the memory consumption of the processes didn't go over the limit.
We're going to have to replace this with something else. Linux CGroups
is one option but the infrastructure I'm right now won't let me use them
so I'm probably going to have to do some sort of nasty polling where
the python script adds up the memory use of the tool and all its child
processes and kills it if the sum is over the memory limit.