Try to fix an issue when running symbooglix and it hits a speculative path without finding bugs during execution. Previously the "failed" attribute would be set to true making it indistinguisable from crashes which is not desirable.
Just changing this attribute is not enough though because if a timeout
wasn't hit then it would be marked as FULLY_EXPLORED by the summary.py
script which is **really** wrong. So I've hacked it so that if
Symbooglix emits the appropriate exit code that the "bound_hit" key is set
to true. This is abusing the meaning of the "bound_hit" a little because the key
was originally meant for hitting a user set bound. However we don't set
one when running Symbooglix normally. So now it also
means effectively that parts of the program were not explored which is
similar to hitting a bound but not quite the same.